refactor: cleared unnecessary divs from the header
