+++ src/ui/_components/molecules/index.ts
@@ -2,5 +2,5 @@ export { SmallFormInput } from './inputs/SmallFormInput'
 export { BigFormInput } from './inputs/BigFormInput'
 export { SelectBox } from './inputs/SelectBox'
 export { FormButton } from './buttons/FormButton'
-export { MainTitle } from './titles/MainTitle'
+export { Header } from './titles/Header'
 export { SectionTitle } from './titles/SectionTitle'
