+++ src/ui/_components/organisms/index.ts
@@ -1,2 +1,1 @@
 export { Form } from './Form'
-export { Header } from './Header'
